+++ Project_1/Project_7_Handout/Source/ModuleEnemies.h
@@ -52,7 +52,7 @@ public:
 	void OnCollision(Collider* c1, Collider* c2) override;
 
 	// Add an enemy into the queue to be spawned later
-	bool AddEnemy(ENEMY_TYPE type, int x, int y);
+	bool AddEnemy(ENEMY_TYPE type, float x, float y);
 
 	// Iterates the queue and checks for camera position
 	void HandleEnemiesSpawn();
